pre loadshedding wip
