+++ tree/node_test.go
@@ -1,46 +1,36 @@
 package tree_test
 
 import (
-	"github.com/sasswart/gin-in-a-can/openapi/media"
-	"github.com/sasswart/gin-in-a-can/openapi/operation"
-	"github.com/sasswart/gin-in-a-can/openapi/path"
-	"github.com/sasswart/gin-in-a-can/openapi/request"
-	"github.com/sasswart/gin-in-a-can/openapi/root"
-	"github.com/sasswart/gin-in-a-can/openapi/schema"
-	"github.com/sasswart/gin-in-a-can/openapi/test"
+	"github.com/sasswart/gin-in-a-can/tree"
 	"testing"
 )
 
-func TestOpenAPI_Dig(t *testing.T) {
-	apiSpec, _ := root.LoadAPISpec(test.AbsOpenAPI)
-	endpoint := test.Dig(apiSpec, test.Endpoint)
-	// TODO check for identity, not just type
-	if _, ok := endpoint.(*path.Item); !ok {
-		t.Errorf("%#v should have been a %T", endpoint, &path.Item{})
-	}
-
-	method := test.Dig(endpoint, test.Method)
-	if _, ok := method.(*operation.Operation); !ok {
-		t.Errorf("%#v should have been a %T", method, &operation.Operation{})
-	}
-
-	reqBody := test.Dig(method, test.ReqBody)
-	if _, ok := reqBody.(*request.Body); !ok {
-		t.Errorf("%#v should have been a %T", reqBody, &request.Body{})
-	}
-
-	mediaType := test.Dig(reqBody, test.MediaType)
-	if _, ok := mediaType.(*media.Type); !ok {
-		t.Errorf("%#v should have been a %T", mediaType, &media.Type{})
-	}
-
-	s := test.Dig(mediaType, test.Schema)
-	if _, ok := s.(*schema.Schema); !ok {
-		t.Errorf("%#v should have been a %T", s, &schema.Schema{})
+func TestOpenAPI_Node_GetAndSetName(t *testing.T) {
+	want := "testname"
+	n := tree.Node{}
+	n.SetName(want)
+	got := n.GetName()
+	if got != want {
+		t.Fail()
 	}
 }
 
+func TestOpenAPI_Node_GetAndSetBasePath(t *testing.T) {
+	want := "testbasepath"
+	n := tree.Node{}
+	n.SetBasePath(want)
+	got := n.GetBasePath()
+	if got != want {
+		t.Fail()
+	}
+}
+
+func TestOpenAPI_Node_GetRef(t *testing.T) {
+	// Should not be implemented for this type
+}
+
 func TestOpenAPI_Node_GetMetadata(t *testing.T) {
+
 	t.Errorf("TODO")
 }
